// src/types/bia.ts
export interface IBiaBodyBenchmark {
  body_score: number;
  physical_age: number;
  body_type	: number;
  result_body_type_description: string;
  recommended_intake_kcal: number;
  ideal_weight: number;
  target_weight: number;
  weight_control: number;
  muscle_control: number;
  fat_control_amount: number;

  
  // 제지방
  lean_body_weight: number;
  lean_body_weight_std_min: number;
  lean_body_weight_std_max: number;
  // 근육량
  muscle_mass: number;
  muscle_mass_std_min: number;
  muscle_mass_std_max: number;
  // 골량
  bone_mass: number;
  bone_mass_std_min: number;
  bone_mass_std_max: number;
  // 세포질량
  body_cell_mass: number;
  body_cell_mass_std_min: number;
  body_cell_mass_std_max: number;
  subcutaneous_fat_mass: number;
  
  // 복부/내장지방 관련
  waist_to_hip_ratio: number;
  waist_to_hip_ratio_std_min: number;
  waist_to_hip_ratio_std_max: number;
  // 세포내수분비
  intracellular_water_volume: number;
  intracellular_water_volume_std_min: number;
  intracellular_water_volume_std_max: number;
  // 표준 체중 대비 체중 비율 
  obesity_percentage: number;
  obesity_percentage_std_min: number;
  obesity_percentage_std_max: number;
  // 피하지방
  subcutaneous_fat_rate: number;
  subcutaneous_fat_rate_std_min: number;
  subcutaneous_fat_rate_std_max: number;
  

}